+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `Facets` field resolver to add `quantity` and make it possible to limit facet values.
+
 ## [1.17.3] - 2020-09-30
 
 ### Fixed